fix: FuncionarioService

+++ ValhallaBebidas.Application/DTOs/MovimentacaoDto.cs
@@ -8,7 +8,7 @@ public class MovimentacaoDto
     public int Id { get; set; }
     public int ProdutoId { get; set; }
     public string NomeProduto { get; set; } = string.Empty;
-    public decimal Quantidade { get; set; }
+    public int Quantidade { get; set; }
     public DirecaoMovimentacao Direcao { get; set; }
     public string Motivo { get; set; } = string.Empty;
     public DateTime Data { get; set; }
@@ -18,7 +18,7 @@ public class MovimentacaoDto
 public class CriarMovimentacaoDto
 {
     public int ProdutoId { get; set; }
-    public decimal Quantidade { get; set; }
+    public int Quantidade { get; set; }
     public DirecaoMovimentacao Direcao { get; set; }
     public string Motivo { get; set; } = string.Empty;
     /* Data e Impacto gerados pelo backend */
@@ -28,7 +28,7 @@ public class AtualizarMovimentacaoDto
 {
     /* Id vem pela URL — não precisa no body */
     public int ProdutoId { get; set; }
-    public decimal Quantidade { get; set; }
+    public int Quantidade { get; set; }
     public DirecaoMovimentacao Direcao { get; set; }
     public string Motivo { get; set; } = string.Empty;
 }
